+++ engine/runtime-with-instruments/src/test/scala/org/enso/interpreter/test/instrument/RuntimeVisualizationsTest.scala
@@ -2954,7 +2954,7 @@ class RuntimeVisualizationsTest
         ConstantsGen.INTEGER,
         Api.MethodPointer(
           warningModuleName.toString,
-          warningTypeName.toString,
+          warningTypeName.toString + ".type",
           "attach"
         )
       ),