+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/Nothing.java
@@ -3,4 +3,9 @@ package org.enso.interpreter.node.expression.builtin;
 import org.enso.interpreter.dsl.BuiltinType;
 
 @BuiltinType(name = "Standard.Base.Nothing.Nothing")
-public class Nothing extends Builtin {}
+public class Nothing extends Builtin {
+  @Override
+  protected boolean containsValues() {
+    return false;
+  }
+}