+++ CHANGELOG.md
@@ -487,6 +487,7 @@
 - [Don't export polyglot symbols][3915]
 - [From/all import must not include module in name resolution][3931]
 - [Vector returns warnings of individual elements][3938]
+- [Enso.getMetaObject, Type.isMetaInstance and Meta.is_a consolidation][3949]
 - [Add executionContext/interrupt API command][3952]
 - [Simplify exception handling for polyglot exceptions][3981]
 
@@ -562,6 +563,7 @@
 [3915]: https://github.com/enso-org/enso/pull/3915
 [3931]: https://github.com/enso-org/enso/pull/3931
 [3938]: https://github.com/enso-org/enso/pull/3938
+[3949]: https://github.com/enso-org/enso/pull/3949
 [3952]: https://github.com/enso-org/enso/pull/3952
 [3981]: https://github.com/enso-org/enso/pull/3981